refactory: 02 단일 파서드 패턴
